Add eslint plugin
